Extracted ZoomCompensate class
